+++ application/i18n/en_GB/misc.php
@@ -1,26 +1,30 @@
 <?php
 
-$lang = array
-(
-  // separators
-	'x_y_separator' => ',',
-	'd_m_s_separator' => ':', // degrees, minutes, seconds for Lat/Long
-  // other language strings
-  'cancel'=>'Cancel',
-  'confirm_delete'=>'Are you sure you want to delete this record?',
-  'confirm_unsubscribe'=>'Are you sure you want to unsubscribe from this notification?',
-  'delete'=>'Delete',
+defined('SYSPATH') or die('No direct access allowed.');
+
+$lang = [
+  // Separators.
+  'x_y_separator' => ',',
+  // Degrees, minutes, seconds for lat/long.
+  'd_m_s_separator' => ':',
+  // Other language strings.
+  'cancel' => 'Cancel',
+  'confirm_delete' => 'Are you sure you want to delete this record?',
+  'confirm_unsubscribe' => 'Are you sure you want to unsubscribe from this notification?',
+  'delete' => 'Delete',
   'indicia_version' => 'Indicia version',
-  'lookup_existing_record'=>'lookup existing record',
-  'new_image'=>'New Image',
-  'next'=>'Next',
-  'save'=>'Save',
-  'unsubscribe'=>'Unsubscribe',
-  // The default notification email texts below can be overridden in
-  // application/config/email.php
-  'notification_subject'=>'Messages from %s',
-  'notification_intro'=>'Here are your latest messages from %s',
-  'then' => 'then',
+  'lookup_existing_record' => 'lookup existing record',
   'model_import' => '%s import',
-  'recheck_verification'=>"This record's verification status needs to be rechecked as it has been edited."
-);
+  'new_image' => 'New Image',
+  'next' => 'Next',
+  'recheck_verification' => "This record's verification status needs to be rechecked as it has been edited.",
+  'save' => 'Save',
+  'then' => 'then',
+  'unsubscribe' => 'Unsubscribe',
+  'untitled' => 'Untitled',
+  // Notification emails.
+  // The default notification email texts below can be overridden in
+  // application/config/email.php.
+  'notification_intro' => 'Here are your latest messages from %s',
+  'notification_subject' => 'Messages from %s',
+];